pyproject.toml
[build-system]
requires = ["hatchling>=1.26", "hatch-vcs"]
build-backend = "hatchling.build"


[project]
name = "orders-of-magnitude"
authors = [
  { name = "Cyrille Praz", email = "cyrraz.code@protonmail.com" },
]
description = "Serve orders of magnitude."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "pint>=0.24",
  "pyyaml>=6",
]

[project.urls]
Homepage = "https://github.com/8-hat/orders-of-magnitude"
"Bug Tracker" = "https://github.com/8-hat/orders-of-magnitude/issues"
Discussions = "https://github.com/8-hat/orders-of-magnitude/discussions"
Changelog = "https://github.com/8-hat/orders-of-magnitude/releases"


[dependency-groups]
test = [
  "pytest >=9",
  "pytest-cov >=7",
]
dev = [
  { include-group = "test" },
]
docs = [
    "markdown>=3.9",
    "mdx-include>=1.4.2",
    "mkdocs-material>=9.1.19",
    "mkdocs>=1.1.2",
    "mkdocstrings-python>=1.18.2",
    "pyyaml>=6.0.1",
]


[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/orders_of_magnitude/_version.py"

[tool.hatch.envs.default]
dependency-groups = ["dev"]

[tool.hatch.envs.hatch-test]
dependency-groups = ["test"]
env-vars.PYTHONWARNDEFAULTENCODING = "1"


[tool.pytest]
minversion = "9.0"
addopts = ["-ra", "--showlocals"]
strict = true
filterwarnings = [
  "error",
]
log_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["orders_of_magnitude"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.10"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "orders_of_magnitude.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true


[tool.ruff]

[tool.ruff.lint]
extend-select = [
  "ARG",    # flake8-unused-arguments
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "DTZ",    # flake8-datetimez
  "EM",     # flake8-errmsg
  "EXE",    # flake8-executable
  "FA",     # flake8-future-annotations
  "FLY",    # flynt
  "FURB",   # refurb
  "G",      # flake8-logging-format
  "I",      # isort
  "ICN",    # flake8-import-conventions
  "ISC",    # flake8-implicit-str-concat
  "LOG",    # flake8-logging
  "NPY",    # NumPy specific rules
  "PD",     # pandas-vet
  "PERF",   # perflint
  "PGH",    # pygrep-hooks
  "PIE",    # flake8-pie
  "PL",     # pylint
  "PT",     # flake8-pytest-style
  "PTH",    # flake8-use-pathlib
  "PYI",    # flake8-pyi
  "Q",      # flake8-quotes
  "RET",    # flake8-return
  "RSE",    # flake8-raise
  "RUF",    # Ruff-specific
  "SIM",    # flake8-simplify
  "SLOT",   # flake8-slots
  "T10",    # flake8-debugger
  "T20",    # flake8-print
  "TC",     # flake8-type-checking
  "TRY",    # tryceratops
  "UP",     # pyupgrade
  "YTT",    # flake8-2020
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
]
# Uncomment if using a _compat.typing backport
# typing-modules = ["orders_of_magnitude._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]
"docs/examples/**" = ["T20"]


[tool.pylint]
py-version = "3.10"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
]
